chore(converter): scope postings to series with chunks in interval
If we happen to converter a one day block out of a 14 day TSDB block we
could run into a situation where we decode and iterate over series that
we know dont have chunks in the particular day we are interested in,
i.e. because of churn.
This change makes it so we just skip those series without looking at any
chunks.

Signed-off-by: Michael Hoffmann <mhoffmann@cloudflare.com>